Move `DataclassOptions` into `types`, beside the class that stores it
`types::class` imported it from `modules::dataclasses`, inverting the
dependency for a field `Class` owns. Define it next to `Class` and have
the module import it back — no behaviour and no serialization change,
since postcard encodes structurally and the dump fingerprints cover
opcodes, static strings and comparison operators, not type paths.

Also reframes the field docstring: the options stand in for the dunders
CPython generates and Monty cannot yet install, rather than reading as
the destination.
